+++ requests_retry_session/__init__.py
@@ -22,6 +22,5 @@
 # OTHER DEALINGS IN THE SOFTWARE.
 #
 
-from .requests_retry_session import requests_retry_session as local_requests_retry_session
-
-requests_retry_session = local_requests_retry_session
+from .requests_retry_session import requests_retry_adapter, requests_retry_session, requests_session
+from .retry_session_manager import RetrySessionManager