+++ requests_retry_session/requests_retry_session.py
@@ -34,23 +34,38 @@ Created on Nov 2, 2020
 @author: jsl
 """
 
-import requests
 # Because we want to support Python 3.6 and 3.9, use old-style type hint syntax
 from typing import Tuple, Union
+
+import requests
 
 from .timeout_http_adapter import TimeoutHTTPAdapter
 from .retry_with_logs import RetryWithLogs
 
-PROTOCOL = 'http'
 
-def requests_retry_session(retries: int = 10,
-                           backoff_factor: float = 0.5,
-                           status_forcelist: Tuple[int, ...] = (500, 502, 503, 504),
-                           connect_timeout: int = 3,
-                           read_timeout: int = 10,
-                           session: Union[None, requests.Session] = None,
-                           protocol: str = PROTOCOL) -> requests.Session:
+DEFAULT_BACKOFF_FACTOR = 0.5
+DEFAULT_CONNECT_TIMEOUT = 3
+DEFAULT_PROTOCOL = 'http'
+DEFAULT_READ_TIMEOUT = 10
+DEFAULT_RETRIES = 10,
+DEFAULT_STATUS_FORCELIST = (500, 502, 503, 504)
+
+
+def requests_session(adapter: requests.adapters.HTTPAdapter,
+                     session: Union[None, requests.Session] = None,
+                     protocol: str = DEFAULT_PROTOCOL) -> requests.Session:
     session = session or requests.Session()
+    # Must mount to http://
+    # Mounting to only http will not work!
+    session.mount(f"{protocol}://", adapter)
+    return session
+
+
+def requests_retry_adapter(retries: int = DEFAULT_RETRIES,
+                           backoff_factor: float = DEFAULT_BACKOFF_FACTOR,
+                           status_forcelist: Tuple[int, ...] = DEFAULT_STATUS_FORCELIST,
+                           connect_timeout: int = DEFAULT_CONNECT_TIMEOUT,
+                           read_timeout: int = DEFAULT_READ_TIMEOUT) -> TimeoutHTTPAdapter:
     retry = RetryWithLogs(
         total=retries,
         read=retries,
@@ -58,8 +73,19 @@ def requests_retry_session(retries: int = 10,
         backoff_factor=backoff_factor,
         status_forcelist=status_forcelist,
     )
-    adapter = TimeoutHTTPAdapter(max_retries=retry, timeout=(connect_timeout, read_timeout))
-    # Must mount to http://
-    # Mounting to only http will not work!
-    session.mount(f"{protocol}://", adapter)
-    return session
+    return TimeoutHTTPAdapter(max_retries=retry, timeout=(connect_timeout, read_timeout))
+
+
+def requests_retry_session(retries: int = DEFAULT_RETRIES,
+                           backoff_factor: float = DEFAULT_BACKOFF_FACTOR,
+                           status_forcelist: Tuple[int, ...] = DEFAULT_STATUS_FORCELIST,
+                           connect_timeout: int = DEFAULT_CONNECT_TIMEOUT,
+                           read_timeout: int = DEFAULT_READ_TIMEOUT,
+                           session: Union[None, requests.Session] = None,
+                           protocol: str = DEFAULT_PROTOCOL) -> requests.Session:
+    adapter = requests_retry_adapter(retries=retries,
+                                     backoff_factor=backoff_factor,
+                                     status_forcelist=status_forcelist,
+                                     connect_timeout=connect_timeout,
+                                     read_timeout=read_timeout)
+    return requests_session(adapter=adapter, session=session, protocol=protocol)